+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     - Add Phasor Reset module.
     - Add Phase Driven Sequencer module.
     - Add Phase Driven Sequencer 32 module.
+    - Add Phasor->Waveforms module.
 - Add Gate Delay module.
 - Add Thomas chaotic attractor to 1-Op Chaos (https://sprott.physics.wisc.edu/chaos/symmetry.htm).
 - Make Mode knobs on 1-Op Chaos, 2-Op Chaos, Chaotic Attactors, and Clocked Noise snap to values and print the current mode.